+++ assets/js/scripts-woocommerce.js
@@ -5,7 +5,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.0.0
+ * @version  1.4.0
  *
  * Contents:
  *
@@ -17,6 +17,8 @@
 
 
 ( function( $ ) {
+
+	'use strict';