+++ library/includes/classes/class-customize-control-multiselect.php
@@ -11,7 +11,7 @@
  *
  * @since    1.0.0
  * @version  2.7.0
- * @version  1.3.0
+ * @version  1.4.0
  */
 class Reykjavik_Customize_Control_Multiselect extends WP_Customize_Control {
 
@@ -72,8 +72,8 @@ class Reykjavik_Customize_Control_Multiselect extends WP_Customize_Control {
 
 			?>
 
-			<span class="customize-control-title"><?php echo $this->label; ?></span>
-			<?php if ( $this->description ) : ?><span class="description customize-control-description"><?php echo $this->description; ?></span><?php endif; ?>
+			<span class="customize-control-title"><?php echo esc_html( $this->label ); ?></span>
+			<?php if ( $this->description ) : ?><span class="description customize-control-description"><?php echo wp_kses_post( $this->description ); ?></span><?php endif; ?>
 
 			<ul>
 			<?php foreach ( $this->choices as $value => $label ) : ?>
@@ -106,8 +106,8 @@ class Reykjavik_Customize_Control_Multiselect extends WP_Customize_Control {
 			?>
 
 			<label>
-				<span class="customize-control-title"><?php echo $this->label; ?></span>
-				<?php if ( $this->description ) : ?><span class="description customize-control-description"><?php echo $this->description; ?></span><?php endif; ?>
+				<span class="customize-control-title"><?php echo esc_html( $this->label ); ?></span>
+				<?php if ( $this->description ) : ?><span class="description customize-control-description"><?php echo wp_kses_post( $this->description ); ?></span><?php endif; ?>
 
 				<select name="<?php echo esc_attr( $this->id ); ?>" multiple="multiple" <?php $this->link(); ?>>
 					<?php foreach ( $this->choices as $value => $label ) : ?>