+++ library/includes/classes/class-customize-control-select.php
@@ -11,7 +11,7 @@
  *
  * @since    1.0.0
  * @version  2.7.0
- * @version  1.3.0
+ * @version  1.4.0
  */
 class Reykjavik_Customize_Control_Select extends WP_Customize_Control {
 
@@ -28,8 +28,8 @@ class Reykjavik_Customize_Control_Select extends WP_Customize_Control {
 				?>
 
 				<label>
-					<span class="customize-control-title"><?php echo $this->label; ?></span>
-					<?php if ( $this->description ) : ?><span class="description customize-control-description"><?php echo $this->description; ?></span><?php endif; ?>
+					<span class="customize-control-title"><?php echo esc_html( $this->label ); ?></span>
+					<?php if ( $this->description ) : ?><span class="description customize-control-description"><?php echo wp_kses_post( $this->description ); ?></span><?php endif; ?>
 
 					<select name="<?php esc_attr( $this->id ); ?>" <?php $this->link(); ?>>
 						<?php