+++ library/init.php
@@ -13,6 +13,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  * @license    GPL-3.0, http://www.gnu.org/licenses/gpl-3.0.html
  * @version    2.7.0
+ * @version    1.4.0
  *
  * Used development prefixes:
  *
@@ -64,50 +65,35 @@
  * 20) Load
  */
 
-	// Core class
+	// Core class.
+	require REYKJAVIK_LIBRARY . 'includes/classes/class-core.php';
 
-		require REYKJAVIK_LIBRARY . 'includes/classes/class-core.php';
+	// Customize (has to be frontend accessible, otherwise it hides the theme settings).
 
-	// Customize (has to be frontend accessible, otherwise it hides the theme settings)
+		// Sanitize class.
+		require REYKJAVIK_LIBRARY . 'includes/classes/class-sanitize.php';
 
-		// Customize class
+		// Customize class.
+		require REYKJAVIK_LIBRARY . 'includes/classes/class-customize.php';
 
-			require REYKJAVIK_LIBRARY . 'includes/classes/class-sanitize.php';
+		// CSS variables generator class.
+		require REYKJAVIK_LIBRARY . 'includes/classes/class-css-variables.php';
 
-		// Customize class
+	// Admin area related functionality.
+	if ( is_admin() ) {
 
-			require REYKJAVIK_LIBRARY . 'includes/classes/class-customize.php';
-
-		// CSS Styles Generator class
-
-			require REYKJAVIK_LIBRARY . 'includes/classes/class-customize-styles.php';
-
-	// Admin
-
-		if ( is_admin() ) {
-
-			// Load the theme welcome page
-
-				locate_template( 'includes/welcome/welcome.php', true );
-
-			// Admin class
-
-				require REYKJAVIK_LIBRARY . 'includes/classes/class-admin.php';
-
-			// Plugins suggestions
-
-				if (
-					(bool) apply_filters( 'wmhook_reykjavik_plugins_suggestion_enabled', true )
-					&& locate_template( 'includes/tgmpa/plugins.php' )
-				) {
-					require REYKJAVIK_LIBRARY . 'includes/vendor/tgmpa/class-tgm-plugin-activation.php';
-					locate_template( 'includes/tgmpa/plugins.php', true );
-				}
-
-			// Child theme generator
-
-				if ( (bool) apply_filters( 'wmhook_reykjavik_child_theme_generator_enabled', false ) ) {
-					require REYKJAVIK_LIBRARY . 'includes/vendor/use-child-theme/class-use-child-theme.php';
-				}
-
+		// Optional theme welcome page.
+		$welcome_page = get_theme_file_path( 'includes/welcome/welcome.php' );
+		if ( file_exists( $welcome_page ) ) {
+			require REYKJAVIK_LIBRARY . 'includes/classes/class-admin.php';
+			require $welcome_page;
 		}
+
+		// Optional plugins suggestions.
+		$plugins_suggestions = get_theme_file_path( 'includes/tgmpa/plugins.php' );
+		if ( (bool) apply_filters( 'wmhook_reykjavik_plugins_suggestion_enabled', file_exists( $plugins_suggestions ) ) ) {
+			require REYKJAVIK_LIBRARY . 'includes/vendor/tgmpa/class-tgm-plugin-activation.php';
+			require $plugins_suggestions;
+		}
+
+	}