+++ library/js/customize-controls.js
@@ -13,8 +13,12 @@
  *
  * @since    2.2.0
  * @version  2.7.0
+ * @version  1.4.0
  */
 ( function( exports, $ ) {
+
+	'use strict';
+
 	$( wp.customize ).on( 'ready', function() {