+++ templates/parts/meta/entry-meta-element-category.php
@@ -6,7 +6,7 @@
  * @copyright  WebMan Design, Oliver Juhas
  *
  * @since    1.0.0
- * @version  1.0.0
+ * @version  1.4.0
  */
 
 
@@ -16,9 +16,9 @@
 // Requirements check
 
 	if (
-			! Reykjavik_Library::is_categorized_blog()
-			|| ! $categories = get_the_category_list( ', ', '', get_the_ID() )
-		) {
+		! Reykjavik_Library::is_categorized_blog()
+		|| ! $categories = get_the_category_list( ', ', '', get_the_ID() )
+	) {
 		return;
 	}
 
@@ -29,5 +29,5 @@
 	<span class="entry-meta-description">
 		<?php echo esc_html_x( 'Categorized in:', 'Post meta info description: categories list.', 'reykjavik' ); ?>
 	</span>
-	<?php echo $categories; ?>
+	<?php echo $categories; /* WPCS: XSS OK. */ ?>
 </span>